build: commit the build artifacts
Commit the build artifact in `dist/` since the
one hosted on the npm registry contains a bug in
it.

By forking the repo under the GitHub organization
we can rebuild the artifacts and commit them so
that we can vendor them in github/github.

Hopefully the package authors will rebuild and
publish so we can switch over from this fork and
just use the package from npm directly.

See this issue I opened up in the original repo[1]

[1]: https://github.com/braintree/restricted-input/issues/43